move framework opts into framework type definitions
