+++ packages/webdriverio/webdriverio-core.d.ts
@@ -51,228 +51,6 @@ declare namespace WebdriverIO {
     interface ServiceWithOption extends Array<string|ServiceOption>{ 0: string; 1: ServiceOption }
     type ServiceEntry = string | HookFunctions | [string, ServiceOption]
 
-    interface MochaOpts {
-        /**
-         * Propagate uncaught errors?
-         */
-        allowUncaught?: boolean;
-        /**
-         * Force done callback or promise?
-         */
-        asyncOnly?: boolean;
-        /**
-         * Bail after first test failure?
-         */
-        bail?: boolean;
-        /**
-         * Check for global variable leaks?
-         */
-        checkLeaks?: boolean;
-        /**
-         * Color TTY output from reporter?
-         */
-        color?: boolean;
-        /**
-         * Delay root suite execution?
-         */
-        delay?: boolean;
-        /**
-         * Show diff on failure?
-         */
-        diff?: boolean;
-        /**
-         * Test filter given string.
-         */
-        fgrep?: string;
-        /**
-         * Tests marked only fail the suite?
-         */
-        forbidOnly?: boolean;
-        /**
-         * Pending tests fail the suite?
-         */
-        forbidPending?: boolean;
-        /**
-         * Full stacktrace upon failure?
-         */
-        fullTrace?: boolean;
-        /**
-         * Variables expected in global scope.
-         */
-        global?: string[];
-        /**
-         * Test filter given regular expression.
-         */
-        grep?: RegExp | string;
-        /**
-         * Enable desktop notifications?
-         */
-        growl?: boolean;
-        /**
-         * Display inline diffs?
-         */
-        inlineDiffs?: boolean;
-        /**
-         * Invert test filter matches?
-         */
-        invert?: boolean;
-        /**
-         * Disable syntax highlighting?
-         */
-        noHighlighting?: boolean;
-        /**
-         * Reporter name or constructor.
-         */
-        reporter?: string | Function;
-        /**
-         * Reporter settings object.
-         */
-        reporterOption?: object;
-        /**
-         * Number of times to retry failed tests.
-         */
-        retries?: number;
-        /**
-         * Slow threshold value.
-         */
-        slow?: number;
-        /**
-         * Timeout threshold value.
-         */
-        timeout?: number | string;
-        /**
-         * Interface name.
-         */
-        ui?: string;
-    }
-
-    interface JasmineNodeOpts {
-        /**
-         * Whether to stop execution of the suite after the first spec failure
-         * Since Jasmine v3.3.0
-         * @default false
-         */
-        failFast?: boolean;
-        /**
-         * Whether to fail the spec if it ran no expectations. By default a spec
-         * that ran no expectations is reported as passed. Setting this to true
-         * will report such spec as a failure.
-         * Since Jasmine v3.5.0
-         * @default false
-         */
-        failSpecWithNoExpectations?: boolean;
-        /**
-         * Whether to cause specs to only have one expectation failure.
-         * Since Jasmine v3.3.0
-         * @default false
-         */
-        oneFailurePerSpec?: boolean;
-        /**
-         * Whether to randomize spec execution order.
-         * Since Jasmine v3.3.0
-         * @default false
-         */
-        random?: boolean;
-        /**
-         * Seed to use as the basis of randomization. Null causes the seed to be
-         * determined randomly at the start of execution.
-         * Since Jasmine v3.3.0
-         * @default null
-         */
-        seed?: Function;
-        /**
-         * Function to use to filter specs.
-         * Since Jasmine v3.3.0
-         * @default true
-         */
-        specFilter?: Function;
-    }
-
-    interface CucumberOpts {
-        /**
-         * Show full backtrace for errors.
-         * @default true
-         */
-        backtrace?: boolean;
-        /**
-         * Require modules prior to requiring any support files.
-         * @default []
-         * @example `['@babel/register']` or `[['@babel/register', { rootMode: 'upward', ignore: ['node_modules'] }]] or [() => { require('ts-node').register({ files: true }) }]`
-         */
-        requireModule?: string[];
-        /**
-         * Treat ambiguous definitions as errors.
-         *
-         * Please note that this is a @wdio/cucumber-framework specific option
-         * and not recognized by cucumber-js itself.
-         * @default false
-         */
-        failAmbiguousDefinitions?: boolean;
-        /**
-         * Abort the run on first failure.
-         * @default false
-         */
-        failFast?: boolean;
-        /**
-         * Treat undefined definitions as warnings.
-         * Please note that this is a @wdio/cucumber-framework specific option and
-         * not recognized by cucumber-js itself.
-         * @default false
-         */
-        ignoreUndefinedDefinitions?: boolean;
-        /**
-         * Only execute the scenarios with name matching the expression (repeatable).
-         * @default []
-         */
-        name?: RegExp[];
-        /**
-         * Specify the profile to use.
-         * @default []
-         */
-        profile?: string[];
-        /**
-         * Require files containing your step definitions before executing features.
-         * You can also specify a glob to your step definitions.
-         * @default []
-         * @example `[path.join(__dirname, 'step-definitions', 'my-steps.js')]`
-         */
-        require?: string[];
-        /**
-         * Specify a custom snippet syntax.
-         */
-        snippetSyntax?: string;
-        /**
-         * Hide step definition snippets for pending steps.
-         * @default true
-         */
-        snippets?: boolean;
-        /**
-         * Hide source uris.
-         * @default true
-         */
-        source?: boolean;
-        /**
-         * Fail if there are any undefined or pending steps
-         * @default false
-         */
-        strict?: boolean
-        /**
-         * Only execute the features or scenarios with tags matching the expression.
-         * Please see the [Cucumber documentation](https://docs.cucumber.io/cucumber/api/#tag-expressions) for more details.
-         */
-        tagExpression?: string;
-        /**
-         * Add cucumber tags to feature or scenario name
-         * @default false
-         */
-        tagsInTitle?: boolean;
-        /**
-         * Timeout in milliseconds for step definitions.
-         * @default 30000
-         */
-        timeout?: number;
-    }
-
     interface Options {
         runner?: string;
         /**
@@ -357,18 +135,6 @@ declare namespace WebdriverIO {
          * Defines the test framework to be used by the WDIO testrunner.
          */
         framework?: string;
-        /**
-         * Mocha specific configurations.
-         */
-        mochaOpts?: MochaOpts;
-        /**
-         * Jasmine specific configurations.
-         */
-        jasmineNodeOpts?: JasmineNodeOpts;
-        /**
-         * Cucumber specific configurations.
-         */
-        cucumberOpts?: CucumberOpts;
         /**
          * List of reporters to use. A reporter can be either a string, or an array of
          * `['reporterName', { <reporter options> }]` where the first element is a string
