+++ tests/typings/sync-cucumber/cucumber.ts
@@ -18,6 +18,8 @@ const hook: WebdriverIO.HookFunctions = {
 
 const config: WebdriverIO.Config = {
     cucumberOpts: {
-        colors: true
+        colors: true,
+        timeout: 123,
+        require: ['123']
     }
 }
