+++ tests/typings/webdriverio-cucumber/cucumber.ts
@@ -1,0 +1,7 @@
+const config: WebdriverIO.Config = {
+    cucumberOpts: {
+        colors: true,
+        timeout: 123,
+        require: ['123']
+    }
+}
