+++ tests/typings/webdriverio-cucumber/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "outDir": "dist",
+    "noImplicitAny": true,
+    "types": [
+      "webdriverio",
+      "@wdio/cucumber-framework"
+    ]
+  }
+}
