+++ tests/typings/webdriverio-mocha/mocha.ts
@@ -7,4 +7,11 @@ const suite: WebdriverIO.Suite = {
     pending: true
 }
 
+const config: WebdriverIO.Config = {
+    mochaOpts: {
+        timeout: 3000,
+        slow: 123
+    }
+}
+
 export default {}
